+++ Packages/com.thenathannator.plasticband/PlasticBand/Devices/FiveFretGuitar/RockBandGuitar/XInputRockBandGuitar.cs
@@ -42,8 +42,7 @@ namespace PlasticBand.Devices
         [InputControl(name = "soloOrange", layout = "MaskButton", format = "USHT", bit = 0, parameters = "mask=0x0140")]
         public ushort buttons;
 
-        // TODO: Define specific ranges for each of the notches
-        [InputControl(name = "pickupSwitch", layout = "Axis", parameters = "scale=true,scaleFactor=4")]
+        [InputControl(name = "pickupSwitch", layout = "RockBandPickupSwitch")]
         public byte pickupSwitch;
 
         public fixed byte unused[5];